Adds metric indicating whether there was a scheduled appointment for an eligible contacts task (Recidiviz/recidiviz-data#54771)
## Description of the change

Sets up new observation and metrics for capturing the % of clients who
are eligible for a task and have a scheduled appointment (thank you
Claude for this assist!)

See query on sandbox for distribution of scheduled appointment hydration
among currently eligible contact tasks:
```
SELECT tasks_contact_type, countif(has_scheduled_contact = "true")/count(*)
FROM `recidiviz-staging.mayukas_54742_observations__person_span.tasks_scheduled_contact_span_materialized`
where current_date between start_date and coalesce(end_date, "9999-12-31") and is_eligible = "true"
group by 1 order by 2 desc
```
which shows that <30% of clients eligible for contacts (among the
contact types that are relevant to JII texts -- namely in-person
home/office contacts) have a scheduled contact recorded in OMS.

There's definitely opportunity to do more to narrow in on the subset of
contact types that are eligible for JII texting, but figured this would
be a good place to start and will facilitate viewing this metric in
looker broken out by compliance task.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f23bf8fcb4b93c7d2db088cf44d4066c39f87341