[US_NC] Create CRR opportunity (Recidiviz/recidiviz-data#54788)
## Description of the change

I created the CRR TES with 3/4 90 consecutive days criteria and an
opportunity record. Specifically:

> 1. 90 consecutive days of active enrollment in an education program OR
> 2. 90 consecutive days of employment, demonstrated by proof of wages
OR
> 3. 90 consecutive days at a facility or institution for medical or
psychological treatment or facility providing rehabilitation,
instruction, recreation, or residence

I uploaded it into a sandbox and it's looking as expected.
```
python -m recidiviz.tools.load_views_to_sandbox \
       --sandbox_dataset_prefix hsalas \
       --state_code_filter US_NC auto \
       --changed_datasets_to_ignore impact_reports,aggregated_metrics,analyst_data\
       --load_changed_views_only
```

## Related issues

Closes Recidiviz/recidiviz-data#54739

## Checklists

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 28936abc1991f3e89835a28f50f31f53648e66e9