[US_IX] Supervision period mapping changes (Recidiviz/recidiviz-data#54736)
## Description of the change

This PR does 3 things:
1. Replaces the Residential string in the mapping to match with the
value sent with idaho. (it includes a different type of space character)
2. Adds 'parole-Non Idaho' to the mappings so these clients don't appear
as a NULL supervision type
3. Adds a few table relationships in mappings. (This was done so that I
could do some testing in the ingest view)

[Validations](https://docs.google.com/document/d/12lmkT1oWnWdjZLOQcDY8PGFtGPKur5y6G0BeocQztVo/edit?tab=t.0)

Successful
[sandbox](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-12-12_09_37_06-10217175869133735916?project=recidiviz-staging)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 56aa830d9e0b521873a5a3fa8e2c90702e6c3e17