[Build] Upgrade Cloud Composer to composer-2.15.3-airflow-2.10.5 (Recidiviz/recidiviz-data#54904)
## Description of the change

Upgrades from Cloud Composer version `composer-2.13.8-airflow-2.10.5` to
`composer-2.15.3-airflow-2.10.5`.

This upgrade addresses security alerts for packages `aiomysql` and
`deepdiff`.

The Airflow metadata DB size in staging is ~1.3 GiB and in prod is ~3.2
GiB. Production is slightly above the recommended 3 GiB threshold, but
should be ok to upgrade. If the upgrade times out, we may need to run
the metadata maintenance DAG first and retry.

**Note**: The next staging and production deploys after this PR merges
will potentially take a long time (as long as an hour) due to the Cloud
Composer upgrade. Platform on-call should be notified before merging.

  ## Type of change

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

  ## Related issues

Closes
https://github.com/Recidiviz/recidiviz-data/security/dependabot/1362
Closes
https://github.com/Recidiviz/recidiviz-data/security/dependabot/1393
Closes
https://github.com/Recidiviz/recidiviz-data/security/dependabot/1396
  Closes Recidiviz/recidiviz-data#51122

  ## Checklists

  ### Development

  **This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

  These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

  ### Code review

  These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: 7fce47a0828c61a4a077bcdd46c7c8d14bdc4cd4