[JII] Add views for JII texting (Recidiviz/recidiviz-data#54931)
## Description of the change

Adds the views described in
[Recidiviz/recidiviz-data#50601](https://github.com/Recidiviz/recidiviz-data/issues/50601) to
`validation_views__jii_texting`

- `invalid_numbers`: returns people where the error from twilio is
related to invalid number
- `text_responses`: product view of folks texting back that aren't stop
related texts along with the PO info.
- `incorrect_contact_reminders`: return people & scheduled contacts
where a given ContactReminderMessageSeries/ContactReminderMessageAttempt
is created but it is not one day before the scheduled contact
- `multiple_welcome_messages`: return people where we've sent more than
one WelcomeMessageSeries/WelcomeMessageAttempt and there's already been
a successful one

**tested in `test_shuff` in production**

https://console.cloud.google.com/bigquery?ws=!1m4!1m3!3m2!1srecidiviz-123!2sshuff_test_validation_views__jii_texting
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-data/issues/50601

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d15fc3bb79ab26c7ca671cfbf098e6aa0df0b1d9