[US_AZ] Add supervisors to state staff periods view (Recidiviz/recidiviz-data#55090)
## Description of the change

We learned that supervisors are not included in the base staff tables,
but only in the base PERSON table. They are referenced in the staff
tables, but need to have their locations pulled in from the PERSON table
to appear in the correct offices and facilities in our tools.

This adds a CTE to the `state_staff_periods` view that grabs the correct
location for each supervisor who is referenced in the base staff tables.
Since this gives us a way to differentiate between staff and
supervisors, I also hydrated `StateStaffRoleSubtype` here in case it is
useful down the line.

In the mapping, I wrapped the `supervisor_periods` and
`caseload_type_periods` hydration in a conditional to ensure that those
are only created for staff, not supervisors. Supervisors will never have
the fields necessary to populate those entities, and if the entity is
created it cannot be left blank.

While I was regenerating fixture files for this view to use real data, I
realized that 3 of the 5 test cases did not have any examples in real
data. Since we established in [this
thread](https://recidiviz.slack.com/archives/C01J3JV7962/p1744835833114439)
that we should only be testing cases we can observe, I removed those
test cases.

**[sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-12-22_07_24_33-15483390614463780701?project=recidiviz-staging)**

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 97ea8af76dd0a7f9c44115838fec9d45fbf4b970